+++ packages/capabilities/types.js
@@ -40,6 +40,7 @@ const BUILTIN_IDS = Object.freeze({
   CAMPAIGN_BUILDER: 'campaign_builder',
   PROPOSAL_GENERATOR: 'proposal_generator',
   MAIL_PACKAGE_GENERATOR: 'mail_package_generator',
+  CAMPAIGN_REVIEW: 'campaign_review',
 });
 
 /**